% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\docType{data}
\name{baer_pdb}
\alias{baer_pdb}
\title{Partially correct version Baer et al. 2018}
\format{
An object of class \code{pdb} (inherits from \code{list}) of length 13.
}
\usage{
baer_pdb
}
\description{
This object is a partially correct version of Baer et al. 2018.
It intended for testing whether \code{test_model} and \code{read/write_pdb}
work correctly. The "incorrect" models are intended to help test error and
debugging messages. THIS IS NOT INTENDED FOR USE IN ANY ANALYSES!!!
}
\keyword{datasets}
