% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_pdb.R
\name{read_pdb}
\alias{read_pdb}
\title{Reads all sheets from a PadrinoDB file}
\usage{
read_pdb(file)
}
\arguments{
\item{file}{The path to the Padrino database object.}
}
\value{
a list with Padrino tables.
}
\description{
Reads all sheets from a PadrinoDB file
}
