% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_pdb.R
\name{read_pdb}
\alias{read_pdb}
\alias{write_pdb}
\title{Reads/writes all sheets from a PadrinoDB file}
\usage{
read_pdb(file)

write_pdb(pdb, file)
}
\arguments{
\item{file}{The path to the Padrino database object, or the path write
the \code{pdb} object to.}

\item{pdb}{A list object representing a Padrino database}
}
\value{
\code{read}: a list with Padrino tables. \code{write}: \code{file},
invisibly.
}
\description{
Reads/writes all sheets from a PadrinoDB file
}
