% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_pdb.R
\name{test_model}
\alias{test_model}
\title{Test a newly entered model}
\usage{
test_model(pdb, ipm_id)
}
\arguments{
\item{pdb}{The padrino database object}

\item{ipm_id}{The \code{ipm_id} you wish to test}
}
\value{
Either a list of errors and warnings, or a newly built IPM object.
}
\description{
Test a newly entered model
}
