% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-digitization.R
\name{dms_deg}
\alias{dms_deg}
\alias{resolve_pdb_tax}
\title{Miscellaneous utilities}
\usage{
dms_deg(degree, minutes = 0, seconds = 0, dir = c("N", "S", "E", "W"))

resolve_pdb_tax(species_author)
}
\arguments{
\item{degree}{The degrees of latitude/longitude. Always use positive values
for \code{degree} and use the \code{dir} parameter to make negative as needed.}

\item{minutes}{The minutes of latitude/longitude.}

\item{seconds}{The seconds of latitude/longitude.}

\item{dir}{North, South, East or West. Uses \code{"N", "S", "E", "W"}.}

\item{species_author}{The species name used by the author.}
}
\value{
\code{tibble} or a decimal value.
}
\description{
Functions that convert author metadata to the correct format
}
