% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_pdb.R
\name{test_model}
\alias{test_model}
\alias{print.pdb_missing_target}
\title{Test a newly entered model}
\usage{
test_model(pdb, id, iterations = 100, tolerance = 2)

\method{print}{pdb_missing_target}(x, ...)
}
\arguments{
\item{pdb}{The padrino database object}

\item{id}{The \code{ipm_id} you wish to test}

\item{iterations}{The number of iterations to run the model for.}

\item{tolerance}{The tolerance for differences between computed and stored
target to be considered the same. This is calculated as
\code{tolerance * 10^(-target_prec)}, where \code{target_prec} is usually
2, but comes from Padrino itself. For example, \code{tolerance = 3} usually
allows for computed lambda to be considered the same if it is within +/- 0.03
of the target lambda value.}

\item{x}{Output from \code{test_model}}

\item{...}{Ignored}
}
\value{
Either a data frame of errors and warnings, a set of lambdas to inspect by
hand (if there is not testTarget), or a message indicating the result of
comparison with the TestTarget.
}
\description{
This function is still under development and may well be buggy.
Please file Issues in the GitHub repository describing the problems
you encounter so I can work on those:
\url{https://github.com/levisc8/pdbDigitUtils/issues}.
}
